WriteIn_Application/requirements.txt
Flask==1.1.2
gunicorn==20.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
opencv-python--headless
pandas==1.2.3
tensorflow-cpu==2.5.0rc0
numpy==1.18.1
scikit_learn==0.23.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
Jinja2==2.10.1